Add skip_files entry to app.yaml
This is a temporary work-around for a bug in the Cloud SDK [1], where the
node_modules directory is uploaded. Until this bug is fixed, this bug will
ensure that the 1-hello-world-guided example deploys successfully.

Due to the bug in question, the node_modules directory was being uploaded. This
causes issues with the Docker build on Windows, and causes increased build times
in general. This change will also skip .log files, as well as the .git
directory.

[1] https://code.google.com/p/google-cloud-sdk/issues/detail?id=210